+++ src/library_browser.js
@@ -507,10 +507,11 @@ var LibraryBrowser = {
 
     // abort and pause-aware versions TODO: build main loop on top of this?
 
-    safeSetTimeout: function(func) {
+    safeSetTimeout: function(func, timeout) {
       // Legacy function, this is used by the SDL2 port so we need to keep it
       // around at least until that is updated.
-      return safeSetTimeout(func);
+      // See https://github.com/libsdl-org/SDL/pull/6304
+      return safeSetTimeout(func, timeout);
     },
     safeRequestAnimationFrame: function(func) {
       {{{ runtimeKeepalivePush() }}}